refactor: start to change the structure of the translator
Remove some unnecessary comments from other libraries and start moving
parts of the translator to separate functions. There is few errors in
the current implementation, as dealing with just the first order of the
best price. Can happen changes in the order book besides the first
order, and the translator should deal with that.