feat and refactor: Make the translator deal with multiple orders by price
Handle multiple orders in different price levels when dealing with rows of the
type BID or ASK.  When the change is various prices above/bellow the current
best- offer or bid, cancel all orders in between and replace the order that was
already there (or inserts a new one). Also, include a break when the time of
the row in beside the valid trading period (16:30) and refactor some codes
after inspect them using profiler.