refactor: change run function and frozen policies
Change the behavior of LeranignAgent when using frozen policies. It does not
take random action anymore, If it is in an unknown state, close out opened
positions and cancel all orders until the environment switch state. Also,
refactor run function to include tags to what kind of test is being performed.